Removed the component framework
Michael Pham and Mr. Fornal decided that the component system was too
complicated for this project. We realized that all we need is an Entity
superclass, and an EntityManager and ResourceManager class. We also need an
EntityBasedGame class that knows about entities, and can invoke their init,
update, and render methods