Completed the test classes. The Entity system works, again!
Removed the component idea. Added a simple test class that contains an Entity that moves based on velocity vector, and a toroidal world (wraps around both edges). The entity simply draws a circle centered on it's location, with a radius sticking out based on the magnitude of its velocity vector.